feat: utilidade para fetch criada
