+++ pages/cadastro-campeonatos.html
@@ -64,6 +64,6 @@
     </main>
     <script type="module" src="../main.js"></script>
     <script type="module" src="../js/utilidades/header.js"></script>
-    <script src="../js/cadastro-campeonatos.js"></script>
+    <script type="module" src="../js/cadastro-campeonatos.js"></script>
 </body>
 </html>
